pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "lm-voting-simulator"
version = "0.1.0"
description = "Synthetic election simulation MVP"
requires-python = ">=3.11"
dependencies = [
  "jinja2",
  "numpy",
  "openai",
  "pandas",
  "pydantic>=2",
  "pyarrow",
  "pyyaml",
  "scikit-learn",
  "typer",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
